Move `LoRAConfig` from `config/__init__.py` to `config/lora.py`
